+++ src/business_logic/transaction/objects/v2/declare_v2.rs
@@ -5,7 +5,7 @@ use crate::{
             TransactionExecutionContext, TransactionExecutionInfo,
         },
         state::state_api::{State, StateReader},
-        state::structs::ExecutionResourcesManager,
+        state::ExecutionResourcesManager,
         transaction::{
             error::TransactionError,
             fee::{calculate_tx_fee, execute_fee_transfer, FeeInfo},
