+++ src/core/syscalls/business_logic_syscall_handler.rs
@@ -28,9 +28,9 @@ use crate::{
             execution_entry_point::ExecutionEntryPoint, CallInfo, CallResult, CallType,
             OrderedEvent, OrderedL2ToL1Message, TransactionExecutionContext,
         },
-        state::structs::ExecutionResourcesManager,
         state::{
             contract_storage_state::ContractStorageState,
+            ExecutionResourcesManager,
             state_api::{State, StateReader},
         },
     },
