+++ tests/complex_contracts/amm_contracts/amm_proxy.rs
@@ -8,7 +8,7 @@ use starknet_rs::{
     business_logic::{
         execution::{CallInfo, CallType},
         state::{cached_state::CachedState, state_api::StateReader},
-        state::{in_memory_state_reader::InMemoryStateReader, structs::ExecutionResourcesManager},
+        state::{in_memory_state_reader::InMemoryStateReader, ExecutionResourcesManager},
     },
     definitions::general_config::StarknetGeneralConfig,
     utils::{calculate_sn_keccak, Address},
