+++ tests/internal_calls.rs
@@ -9,7 +9,7 @@ use starknet_rs::{
             execution_entry_point::ExecutionEntryPoint, CallType, TransactionExecutionContext,
         },
         state::{cached_state::CachedState, state_cache::StorageEntry},
-        state::{in_memory_state_reader::InMemoryStateReader, structs::ExecutionResourcesManager},
+        state::{in_memory_state_reader::InMemoryStateReader, ExecutionResourcesManager},
     },
     definitions::{constants::TRANSACTION_VERSION, general_config::StarknetGeneralConfig},
     services::api::contract_classes::deprecated_contract_class::ContractClass,
