+++ tests/syscalls_errors.rs
@@ -11,7 +11,7 @@ use starknet_rs::{
             cached_state::{CachedState, ContractClassCache},
             state_api::State,
         },
-        state::{in_memory_state_reader::InMemoryStateReader, structs::ExecutionResourcesManager},
+        state::{in_memory_state_reader::InMemoryStateReader, ExecutionResourcesManager},
     },
     core::errors::state_errors::StateError,
     definitions::{constants::TRANSACTION_VERSION, general_config::StarknetGeneralConfig},
